# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateMenuRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        maxLength: 180
        type: string
      price:
        minimum: 0
        type: number
    required:
    - name
    - price
    type: object
...
  dto.ToggleMenuAvailableRequest:
    properties:
      is_available:
        type: boolean
    required:
    - is_available
    type: object
...
  dto.UpdateMenuRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        maxLength: 180
        type: string
      price:
        minimum: 0
        type: number
    type: object
...
info:
  contact: {}
  description: Backend API Service untuk platform SaaS POS & Self-Order UMKM kuliner.
  title: saas_gangsta Backend API
  version: "1.0"
paths:
...
  /menus:
    get:
      description: Mengambil daftar menu per tenant (mendukung filter category_id
        dan is_available)
      parameters:
      - description: Filter berdasarkan Category ID
        in: query
        name: category_id
        type: string
      - description: Filter berdasarkan ketersediaan (true/false)
        in: query
        name: is_available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Menus
      tags:
      - Menu Management
    post:
      consumes:
      - application/json
      description: Mendaftarkan menu baru
      parameters:
      - description: Payload Create Menu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Menu
      tags:
      - Menu Management
  /menus/{id}:
    delete:
      description: Menghapus menu (soft delete)
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Soft Delete Menu
      tags:
      - Menu Management
    get:
      description: Mengambil detail satu menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detail Menu
      tags:
      - Menu Management
    put:
      consumes:
      - application/json
      description: Memperbarui informasi menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Update Menu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Menu
      tags:
      - Menu Management
  /menus/{id}/toggle-available:
    patch:
      consumes:
      - application/json
      description: Mengubah status is_available pada menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Toggle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ToggleMenuAvailableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Toggle menu availability
      tags:
      - Menu Management
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"